PAE-543 adjust logic for running seed data cleanup in prod (#396)
- changes how the code determines if its running in production from config.get('isProduction') to config.get('cdpEnvironment') === 'prod'
- re-instates the epr-organisation fixtures (if they're not already in the epr-organisation collection)
- adds a "dry run" option to the code that cleans up seed data (on by default)

+++ src/common/helpers/collections/create-update.js
@@ -35,6 +35,15 @@ const COLLECTION_REGISTRATION = 'registration'
 const COLLECTION_ACCREDITATION = 'accreditation'
 const COLLECTION_EPR_ORGANISATIONS = 'epr-organisations'
 
+const eprOrganisationFixturesIds = [
+  eprOrganisation1,
+  eprOrganisation2,
+  eprOrganisation3,
+  eprOrganisation4
+]
+  .map((record) => record.id)
+  .map(ObjectId.createFromHexString)
+
 /**
  * @import {Db} from 'mongodb'
  */
@@ -91,13 +100,16 @@ export async function createIndexes(db) {
  */
 export async function createSeedData(db, isProduction) {
   if (!isProduction()) {
-    logger.info({ message: 'Creating seed data' })
+    logger.info({ message: 'Create seed data: start' })
 
     const organisationDocCount = await db
       .collection(COLLECTION_ORGANISATION)
       .countDocuments()
 
     if (organisationDocCount === 0) {
+      logger.info({
+        message: 'Create seed data: inserting org/reg/acc fixtures'
+      })
       const organisationAnswers = extractAnswers(organisationFixture)
 
       const { insertedIds } = await db
@@ -132,11 +144,15 @@ export async function createSeedData(db, isProduction) {
       ])
     }
 
-    const eprOrganisationDocCount = await db
+    const eprOrganisationFixturesDocs = await db
       .collection(COLLECTION_EPR_ORGANISATIONS)
-      .countDocuments()
+      .find({ _id: { $in: eprOrganisationFixturesIds } })
+      .toArray()
 
-    if (eprOrganisationDocCount === 0) {
+    if (eprOrganisationFixturesDocs.length === 0) {
+      logger.info({
+        message: 'Create seed data: inserting epr-organisation fixtures'
+      })
       await db
         .collection(COLLECTION_EPR_ORGANISATIONS)
         .insertMany([
@@ -149,15 +165,26 @@ export async function createSeedData(db, isProduction) {
   }
 }
 
-export async function cleanupSeedData(db, isProduction) {
+export async function cleanupSeedData(db, { isProduction, isDryRun }) {
+  logger.info({
+    message: `Seed data clean up: requested, isProduction-${isProduction()}, dry run-${isDryRun()}`
+  })
   if (isProduction()) {
+    logger.info({ message: 'Seed data clean up: start' })
     const deleteDocuments = async (collectionName, ids) => {
-      const result = await db
-        .collection(collectionName)
-        .deleteMany({ _id: { $in: ids } })
-      logger.info({
-        message: `Seed data clean up: deleted ${result.deletedCount} documents from ${collectionName} collection`
-      })
+      if (isDryRun()) {
+        const idStrings = ids.map((id) => id.toString())
+        logger.info({
+          message: `Seed data clean up: dry run delete documents ${idStrings} from ${collectionName} collection`
+        })
+      } else {
+        const result = await db
+          .collection(collectionName)
+          .deleteMany({ _id: { $in: ids } })
+        logger.info({
+          message: `Seed data clean up: deleted ${result.deletedCount} documents from ${collectionName} collection`
+        })
+      }
     }
 
     const findAndDeleteOne = async (collectionName, query) => {
@@ -190,15 +217,9 @@ export async function cleanupSeedData(db, isProduction) {
         accreditationFixture.data.main.MyWHms // system reference
     })
 
-    const eprOrganisationIds = [
-      eprOrganisation1,
-      eprOrganisation2,
-      eprOrganisation3,
-      eprOrganisation4
-    ]
-      .map((record) => record.id)
-      .map(ObjectId.createFromHexString)
-
-    await deleteDocuments(COLLECTION_EPR_ORGANISATIONS, eprOrganisationIds)
+    await deleteDocuments(
+      COLLECTION_EPR_ORGANISATIONS,
+      eprOrganisationFixturesIds
+    )
   }
 }
